refactor of presidential lists - moves url from presidential-staff to presidential-lists - makes use of toggleUrl function for API calls - implements API call to new backend endpoint - changes title to use party alias instead of party name - adds contitionals to VPs in case party does not have them
