+++ components/Steps/PresidentialList/styles.js
