--- pages/presidential-staff/[partyName]/index.js
@@ -1,11 +1,0 @@
-import dynamic from 'next/dynamic';
-
-const PresidentialStaff = dynamic(
-  () => import('components/Steps/PresidentialStaff'),
-  {
-    ssr: false,
-  },
-);
-
-const nossrPresidentialStaff = () => <PresidentialStaff />;
-export default nossrPresidentialStaff;
